+++ p2panda-net/src/test_utils.rs
@@ -341,7 +341,7 @@ pub fn create_operation(
         signature: None,
         payload_size: body.size(),
         payload_hash: Some(body.hash()),
-        timestamp,
+        timestamp: timestamp.into(),
         seq_num,
         backlink,
         extensions: (),